[ACR] Use freertos types in logging

+++ source/azure_iot_json_reader.c
@@ -33,7 +33,7 @@ AzureIoTResult_t AzureIoTJSONReader_Init( AzureIoTJSONReader_t * pxReader,
 
         if( az_result_failed( xCoreResult = az_json_reader_init( &pxReader->_internal.xCoreReader, xJSONSpan, NULL ) ) )
         {
-            AZLogError( ( "Could not initialize the JSON reader: core error=0x%08x", ( unsigned int ) xCoreResult ) );
+            AZLogError( ( "Could not initialize the JSON reader: core error=0x%08x", ( uint16_t ) xCoreResult ) );
             xResult = AzureIoT_TranslateCoreError( xCoreResult );
         }
         else
@@ -60,7 +60,7 @@ AzureIoTResult_t AzureIoTJSONReader_NextToken( AzureIoTJSONReader_t * pxReader )
     {
         if( az_result_failed( xCoreResult = az_json_reader_next_token( &pxReader->_internal.xCoreReader ) ) )
         {
-            AZLogError( ( "Could not get next JSON token: core error=0x%08x", ( unsigned int ) xCoreResult ) );
+            AZLogError( ( "Could not get next JSON token: core error=0x%08x", ( uint16_t ) xCoreResult ) );
             xResult = AzureIoT_TranslateCoreError( xCoreResult );
         }
         else
@@ -86,7 +86,7 @@ AzureIoTResult_t AzureIoTJSONReader_SkipChildren( AzureIoTJSONReader_t * pxReade
     {
         if( az_result_failed( xCoreResult = az_json_reader_skip_children( &pxReader->_internal.xCoreReader ) ) )
         {
-            AZLogError( ( "Could not skip children in JSON: core error=0x%08x", ( unsigned int ) xCoreResult ) );
+            AZLogError( ( "Could not skip children in JSON: core error=0x%08x", ( uint16_t ) xCoreResult ) );
             xResult = AzureIoT_TranslateCoreError( xCoreResult );
         }
         else
@@ -114,7 +114,7 @@ AzureIoTResult_t AzureIoTJSONReader_GetTokenBool( AzureIoTJSONReader_t * pxReade
     {
         if( az_result_failed( xCoreResult = az_json_token_get_boolean( &pxReader->_internal.xCoreReader.token, pxValue ) ) )
         {
-            AZLogError( ( "Could not get boolean in JSON: core error=0x%08x", ( unsigned int ) xCoreResult ) );
+            AZLogError( ( "Could not get boolean in JSON: core error=0x%08x", ( uint16_t ) xCoreResult ) );
             xResult = AzureIoT_TranslateCoreError( xCoreResult );
         }
         else
@@ -142,7 +142,7 @@ AzureIoTResult_t AzureIoTJSONReader_GetTokenInt32( AzureIoTJSONReader_t * pxRead
     {
         if( az_result_failed( xCoreResult = az_json_token_get_int32( &pxReader->_internal.xCoreReader.token, plValue ) ) )
         {
-            AZLogError( ( "Could not get int32_t in JSON: core error=0x%08x", ( unsigned int ) xCoreResult ) );
+            AZLogError( ( "Could not get int32_t in JSON: core error=0x%08x", ( uint16_t ) xCoreResult ) );
             xResult = AzureIoT_TranslateCoreError( xCoreResult );
         }
         else
@@ -170,7 +170,7 @@ AzureIoTResult_t AzureIoTJSONReader_GetTokenDouble( AzureIoTJSONReader_t * pxRea
     {
         if( az_result_failed( xCoreResult = az_json_token_get_double( &pxReader->_internal.xCoreReader.token, pxValue ) ) )
         {
-            AZLogError( ( "Could not get double in JSON: core error=0x%08x", ( unsigned int ) xCoreResult ) );
+            AZLogError( ( "Could not get double in JSON: core error=0x%08x", ( uint16_t ) xCoreResult ) );
             xResult = AzureIoT_TranslateCoreError( xCoreResult );
         }
         else
@@ -201,7 +201,7 @@ AzureIoTResult_t AzureIoTJSONReader_GetTokenString( AzureIoTJSONReader_t * pxRea
         if( az_result_failed( xCoreResult = az_json_token_get_string( &pxReader->_internal.xCoreReader.token,
                                                                       ( char * ) pucBuffer, ( int32_t ) ulBufferSize, ( int32_t * ) pusBytesCopied ) ) )
         {
-            AZLogError( ( "Could not get string in JSON: core error=0x%08x", ( unsigned int ) xCoreResult ) );
+            AZLogError( ( "Could not get string in JSON: core error=0x%08x", ( uint16_t ) xCoreResult ) );
             xResult = AzureIoT_TranslateCoreError( xCoreResult );
         }
         else
